refactor: extract storage logic into FilmStorage and UserStorage with InMemory implementations (@Component)
